WIP! anv: Implement CPU-side BVH building with embree
This commit also adds an environment variable to force a CPU-side build
in vkCmdBuildAccelerationStructuresKHR.  This is useful for getting more
tests working because most tests assume GPU-side builds.  Also, when
running in simulation, a CPU-side build with Embree is likely to be much
faster than a GPU-side build with simulated compute shaders.  This, of
course, assumes that the data is already in the memory at the time that
vkCmdBuildAccelerationStructureKHR is called.  However, this is likely
true for most CTS tests and we can get more stuff running this way.

TODO: Take advantage of deferred operations